+++ src/main/java/http/cookie/Cookie.java
@@ -1,4 +1,4 @@
-package http.response;
+package http.cookie;
 
 import java.util.Objects;
