+++ src/main/java/webserver/RequestHandler.java
@@ -4,7 +4,7 @@ import http.request.Request;
 import http.response.Response;
 import http.response.ResponseHeader;
 import http.response.StatusLine;
-import http.utils.RequestFactory;
+import http.request.RequestFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import webserver.controller.Controller;
