+++ src/main/java/webserver/support/CookieParser.java
@@ -2,7 +2,7 @@ package webserver.support;
 
 import com.github.jknack.handlebars.internal.lang3.StringUtils;
 import com.google.common.collect.Lists;
-import http.response.Cookie;
+import http.cookie.Cookie;
 
 import java.util.Collections;
 import java.util.List;
