+++ src/test/java/http/request/RequestTest.java
@@ -1,6 +1,5 @@
 package http.request;
 
-import http.utils.RequestFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
