+++ src/test/java/webserver/controller/UserListControllerTest.java
@@ -1,7 +1,7 @@
 package webserver.controller;
 
 import db.DataBase;
-import http.response.Cookies;
+import http.cookie.Cookies;
 import model.User;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
